Separated Interface Class model into separate class file.
