+++ Virtual_EDW/Classes/ClassInterfaceHandling.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Virtual_Data_Warehouse;
+using Virtual_Data_Warehouse.Classes;
 
 namespace Virtual_Data_Warehouse
 {
